+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## 4.0
 
+ - TODO: Prevent teleporting into an arena
+
  - Updated to Spigot API 1.14, will no longer work on previous versions. May also require Magic 8.1 or higher.
  - Add item_wear arena property, to turn off items and armor getting worn in a battle
  - Add allow_consuming arena property
  - Fix arena config changes not saving when setting some properties
+ - Add leaderboard_sign_type parameter to arenas
 
 ## 3.4